improvement: add `$ne` operator support to `reactiveQueries`
This query is useful for when you want to listen for when a field is not
`undefined`. For example:
`Publications.getCollection('meetingtemplates').find({ copyOf: { $ne:
undefined } });`

This enables that and adds tests. Unfortunately, to add tests correctly,
we also needed to upgrade babel since jest supports it really nicely
(that's what the new babelrc is for).