feat(localcollection): allow supressing removal warnings
For a long time we've dealt with unactionable errors when the
`_onRemoved` function is triggered for `LocalCollection`s in cases where
we intentionally, optimistically removed a document from the collection.

This allows consumers of `publication-client` the ability to opt-in to
warning supression. This achieves the goal of warning about potential
issues with document removal, but allows one to investigate, determine
if there's an issue, and opt-out of noisy warnings.