fix: remove unnecessary errors
For a long time, we've thrown errors when a document was changed or
removed without being added. These errors have almost never been
actionable and have not revealed a bug to my knowledge. As we move to
avoid bootstrapping when possible, we're intentionally sending change
events without specifically adding the item.

This also helps avoid cumbersome tracking of whether or not the client
has received an object in the publications server.